Remove remaining traces of "plugin"
