+++ index.js
@@ -2,9 +2,6 @@ import React from "react";
 import { render } from "react-dom";
 
 import KintoAdmin from "./src";
-import * as signoffPlugin from "./src/plugins/signoff";
-
-const corePlugins = [signoffPlugin];
 
 const settings = {
   // Enforce a hard limit of 200 max items per page for records and history
@@ -13,7 +10,4 @@ const settings = {
   maxPerPage: 200,
 };
 
-render(
-  <KintoAdmin plugins={corePlugins} settings={settings} />,
-  document.getElementById("app")
-);
+render(<KintoAdmin settings={settings} />, document.getElementById("app"));
