+++ src/routes.tsx
@@ -1,49 +1,14 @@
 import * as React from "react";
-import { Component, PureComponent } from "react";
+import { PureComponent } from "react";
 import { Route, Switch } from "react-router-dom";
 import type { RouteComponentProps } from "react-router-dom";
-import { utils as formUtils } from "kinto-admin-form";
 import type { Dispatch } from "redux";
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
 import * as RouteActions from "./actions/route";
 
-import { isObject } from "./utils";
 import App from "./containers/App";
-import Sidebar from "./containers/Sidebar";
-import Notifications from "./containers/Notifications";
-import CollectionRecordsPage from "./containers/collection/CollectionRecordsPage";
-
-function registerPluginsComponentHooks(
-  PageContainer,
-  plugins
-): React.ComponentType {
-  // Extract the container wrapped component (see react-redux connect() API)
-  const { WrappedComponent } = PageContainer;
-  // By convention, the hook namespace is the wrapped component name
-  const namespace = WrappedComponent.displayName;
-  if (!namespace) {
-    throw new Error("can't happen -- component with no display name");
-  }
-  // Retrieve all the hooks if any
-  const hooks = plugins.map(plugin => plugin.hooks).filter(isObject);
-  // Merge all the hooks together, recursively grouped by namespaces
-  const mergedHooks = hooks.reduce((acc, hookObject) => {
-    return formUtils.mergeObjects(acc, hookObject, true);
-  }, {});
-  // Wrap the root component, augmenting its props with the plugin hooks for it.
-  return class extends Component {
-    render() {
-      return (
-        <PageContainer
-          {...this.props}
-          pluginHooks={mergedHooks[namespace] || {}}
-        />
-      );
-    }
-  };
-}
 
 type ComponentWrapperProps = RouteComponentProps & {
   component: React.ComponentType;
@@ -149,29 +114,10 @@ function mapDispatchToProps(dispatch: Dispatch): {
 
 export const CreateRoute = connect(null, mapDispatchToProps)(routeCreator);
 
-export default function getRoutes(store: Object, plugins: Object[] = []) {
-  const hookedSidebar = registerPluginsComponentHooks(Sidebar, plugins);
-  const hookedNotifications = registerPluginsComponentHooks(
-    Notifications,
-    plugins
-  );
-  const hookedCollectionRecords = registerPluginsComponentHooks(
-    CollectionRecordsPage,
-    plugins
-  );
-
+export default function getRoutes(store: Object) {
   return (
     <Switch>
-      <Route
-        path="/"
-        render={props => (
-          <App
-            sidebar={hookedSidebar}
-            notifications={hookedNotifications}
-            collectionRecords={hookedCollectionRecords}
-          />
-        )}
-      />
+      <Route path="/" component={App} />
     </Switch>
   );
 }
